Changes that conflict with Round2 Issues.
Each of the 190-Round2 issues in the issue tracker add links to the README.md file.  This commit, when merged into main will introduce changes that conflict with all of the changes requested by those issues. It should be merged after students synch with main and as or just after they submit their PR's for the 190-Round2 issues.